# Improve `Flipper#enabled?` Performance