--- a/lib/flipper/gate_values.rb
+++ b/lib/flipper/gate_values.rb
@@ -3,16 +3,6 @@ require 'flipper/typecast'
 
 module Flipper
   class GateValues
-    # Private: Array of instance variables that are readable through the []
-    # instance method.
-    LegitIvars = {
-      'boolean' => '@boolean',
-      'actors' => '@actors',
-      'groups' => '@groups',
-      'percentage_of_time' => '@percentage_of_time',
-      'percentage_of_actors' => '@percentage_of_actors',
-    }.freeze
-
     attr_reader :boolean
     attr_reader :actors
     attr_reader :groups
@@ -25,12 +15,6 @@ module Flipper
       @groups = Typecast.to_set(adapter_values[:groups])
       @percentage_of_actors = Typecast.to_percentage(adapter_values[:percentage_of_actors])
       @percentage_of_time = Typecast.to_percentage(adapter_values[:percentage_of_time])
-    end
-
-    def [](key)
-      if ivar = LegitIvars[key.to_s]
-        instance_variable_get(ivar)
-      end
     end
 
     def eql?(other)