--- a/lib/flipper/gates/boolean.rb
+++ b/lib/flipper/gates/boolean.rb
@@ -24,7 +24,7 @@ module Flipper
       # Returns true if explicitly set to true, false if explicitly set to false
       # or nil if not explicitly set.
       def open?(context)
-        context.values.send(key)
+        context.values.boolean
       end
 
       def wrap(thing)