--- a/lib/flipper/gates/group.rb
+++ b/lib/flipper/gates/group.rb
@@ -23,7 +23,7 @@ module Flipper
       #
       # Returns true if gate open for thing, false if not.
       def open?(context)
-        value = context.values[key]
+        value = context.values.send(key)
         if context.thing.nil?
           false
         else