--- a/lib/flipper/gates/percentage_of_actors.rb
+++ b/lib/flipper/gates/percentage_of_actors.rb
@@ -21,15 +21,18 @@ module Flipper
         value > 0
       end
 
+      # Private: this constant is used to support up to 3 decimal places
+      # in percentages.
       SCALING_FACTOR = 1_000
+      private_constant :SCALING_FACTOR
 
       # Internal: Checks if the gate is open for a thing.
       #
       # Returns true if gate open for thing, false if not.
       def open?(context)
         return false if context.thing.nil?
+
         id = "#{context.feature_name}#{context.thing.value}"
-        # this is to support up to 3 decimal places in percentages
         Zlib.crc32(id) % (100 * SCALING_FACTOR) < context.values.percentage_of_actors * SCALING_FACTOR
       end