--- a/lib/flipper/typecast.rb
+++ b/lib/flipper/typecast.rb
@@ -21,11 +21,9 @@ module Flipper
     # Returns an Integer representation of the value.
     # Raises ArgumentError if conversion is not possible.
     def self.to_integer(value)
-      if value.respond_to?(:to_i)
-        value.to_i
-      else
-        raise ArgumentError, "#{value.inspect} cannot be converted to an integer"
-      end
+      value.to_i
+    rescue NoMethodError
+      raise ArgumentError, "#{value.inspect} cannot be converted to an integer"
     end
 
     # Internal: Convert value to a float.
@@ -33,11 +31,9 @@ module Flipper
     # Returns a Float representation of the value.
     # Raises ArgumentError if conversion is not possible.
     def self.to_float(value)
-      if value.respond_to?(:to_f)
-        value.to_f
-      else
-        raise ArgumentError, "#{value.inspect} cannot be converted to a float"
-      end
+      value.to_f
+    rescue NoMethodError
+      raise ArgumentError, "#{value.inspect} cannot be converted to a float"
     end
 
     # Internal: Convert value to a percentage.
@@ -45,11 +41,11 @@ module Flipper
     # Returns a Integer or Float representation of the value.
     # Raises ArgumentError if conversion is not possible.
     def self.to_percentage(value)
-      if value.to_s.include?('.'.freeze)
-        to_float(value)
-      else
-        to_integer(value)
-      end
+      result_to_f = value.to_f
+      result_to_i = result_to_f.to_i
+      result_to_f == result_to_i ? result_to_i : result_to_f
+    rescue NoMethodError
+      raise ArgumentError, "#{value.inspect} cannot be converted to a percentage"
     end
 
     # Internal: Convert value to a set.