@bkeepers and I paired on some Flipper#enabled? performance. 

When we started:

```
          with actor    164.802k (± 0.6%) i/s -    832.116k in   5.049373s
       without actor    259.654k (± 0.3%) i/s -      1.307M in   5.034628s
```

When we finished:

```
          with actor    255.295k (± 0.9%) i/s -      1.300M in   5.092099s
       without actor    356.474k (± 0.9%) i/s -      1.787M in   5.013634s
```

~55% for with actor. ~37% without an actor.